refactor: class 별로 파일 분할, calendarManager 클래스 상속으로 리팩토링
